+++ js/getRandom.js
@@ -1,0 +1,3 @@
+export function getRandom(a, b) {
+    return Math.floor(Math.random() * (b - a + 1)) + a;
+  }
