+++ js/isCorrect.js
@@ -1,0 +1,3 @@
+export function isCorrect(str, maxLen) {
+  return str.length <= maxLen;
+}
