+++ js/photoObject.js
@@ -1,0 +1,10 @@
+export function photo(id, url, description, likes, comments) {
+  this.id = id;
+  this.url = url;
+  this.description = description;
+  this.likes = likes;
+  this.comments = comments;
+}
+export const descriptions = ['1', '2', '3', '4', '5', '6', '7', '8', '9',
+'10','11', '12', '13', '14', '15','16', '17', '18',
+'19', '20', '21', '22', '23', '24', '25',];
