+++ package.json
@@ -2,6 +2,7 @@
   "name": "kekstagram",
   "version": "26.0.0",
   "private": true,
+  "type": "module",
   "description": "Личный проект «Кекстаграм (простой)» от HTML Academy",
   "repository": {
     "type": "git",
@@ -25,4 +26,6 @@
     "node": "16.x",
     "npm": "8.x"
   }
+
+
 }
